Begin making common parameters
