+++ src/mx_bluesky/common/device_setup_plans/read_hardware_for_setup.py
@@ -1,7 +1,7 @@
 import bluesky.plan_stubs as bps
 from dodal.devices.eiger import EigerDetector
 
-from mx_bluesky.common.parameters.constants import MxConstants
+from mx_bluesky.common.parameters.constants import MxDefaultConstants
 
 
 def read_hardware_for_zocalo(detector: EigerDetector):
@@ -9,6 +9,6 @@ def read_hardware_for_zocalo(detector: EigerDetector):
     If the RunEngine is subscribed to the ZocaloCallback, this plan will also trigger zocalo.
     A bluesky run must be open to use this plan
     """
-    yield from bps.create(name=MxConstants.DESCRIPTORS.ZOCALO_HW_READ)
+    yield from bps.create(name=MxDefaultConstants.DESCRIPTORS.ZOCALO_HW_READ)
     yield from bps.read(detector.odin.file_writer.id)  # type: ignore # See: https://github.com/bluesky/bluesky/issues/1809
     yield from bps.save()
